requirements.txt
streamlit
youtube_transcript_api
gtts
python-dotenv
openai